// seungjae20260713/api-server/src/sample/types.ts
export interface CompanyProfile {
	ticker: string;
	name: string;
	market: Market;
	currency: Currency;
	description: string;
	industry: string;
	sector: string;
	country: string;
	mainBusiness: string;
	competitors: string[];
}